Update mod_jhu_death_rate.R
That formula (death/diagnosed) underestimates the correct case fatality rate: "The CFR number during the course of an outbreak with a high daily increase and long resolution time would be substantially lower than the final CFR" in https://en.wikipedia.org/wiki/Case_fatality_rate. Some people propose to do the ratio on "resolved cases" - that is all cases that have resolved (death or recovery) to minimize the risk of underestimation due to the fact that some of the cases in progress will result, unfortunately, in a fatality.